+++ Source/KIS_Shared.cs
@@ -1,4 +1,5 @@
 ﻿using KISAPIv1;
+using KSPDev.ConfigUtils;
 using KSPDev.LogUtils;
 using KSPDev.ProcessingUtils;
 using KSPDev.PartUtils;
@@ -55,25 +56,6 @@ public static class KIS_Shared {
 
   public static void PlaySoundAtPoint(string soundPath, Vector3 position) {
     AudioSource.PlayClipAtPoint(GameDatabase.Instance.GetAudioClip(soundPath), position);
-  }
-
-  public static bool createFXSound(Part part, FXGroup group, string sndPath, bool loop,
-                                   float maxDistance = 30f) {
-    group.audio = part.gameObject.AddComponent<AudioSource>();
-    group.audio.volume = GameSettings.SHIP_VOLUME;
-    group.audio.rolloffMode = AudioRolloffMode.Linear;
-    group.audio.dopplerLevel = 0f;
-    group.audio.spatialBlend = 1f;
-    group.audio.maxDistance = maxDistance;
-    group.audio.loop = loop;
-    group.audio.playOnAwake = false;
-    if (GameDatabase.Instance.ExistsAudioClip(sndPath)) {
-      group.audio.clip = GameDatabase.Instance.GetAudioClip(sndPath);
-      return true;
-    } else {
-      DebugEx.Error("Sound not found in the game database !");
-      return false;
-    }
   }
 
   /// <summary>
@@ -152,27 +134,6 @@ public static class KIS_Shared {
         part.vessel.vesselName = sourceVessel.vesselName + " 1";
       }
     }
-  }
-
-  public static ConfigNode vesselSnapshot(Vessel vessel) {
-    ProtoVessel snapshot = new ProtoVessel(vessel);
-    ConfigNode node = new ConfigNode("VESSEL");
-    snapshot.Save(node);
-    return node;
-  }
-
-  public static Collider GetEvaCollider(Vessel evaVessel, string colliderName) {
-    KerbalEVA kerbalEva = evaVessel.rootPart.gameObject.GetComponent<KerbalEVA>();
-    Collider evaCollider = null;
-    if (kerbalEva) {
-      foreach (var col in kerbalEva.characterColliders) {
-        if (col.name == colliderName) {
-          evaCollider = col;
-          break;
-        }
-      }
-    }
-    return evaCollider;
   }
 
   public static Part CreatePart(AvailablePart avPart, Vector3 position, Quaternion rotation,
@@ -255,6 +216,8 @@ public static class KIS_Shared {
     newPart.transform.position = position;
     newPart.transform.rotation = rotation;
     newPart.missionID = fromPart.missionID;
+    newPart.launchID = ConfigAccessor.GetValueByPath<uint>(partConfig, "launchID")
+        ?? fromPart.launchID;
     newPart.UpdateOrgPosAndRot(newPart.vessel.rootPart);
 
     if (coupleToPart != null) {
@@ -513,87 +476,6 @@ public static class KIS_Shared {
       node.icon.transform.localPosition = Vector3.zero;
       node.icon.transform.localRotation = Quaternion.identity;
     }
-  }
-
-  public static void EditField(string label, ref bool value, int maxLenght = 50) {
-    value = GUILayout.Toggle(value, label);
-  }
-
-  public static Dictionary<string, string> editFields = new Dictionary<string, string>();
-
-  public static bool EditField(string label, ref Vector3 value, int maxLenght = 50) {
-    bool btnPress = false;
-    if (!editFields.ContainsKey(label + "x")) {
-      editFields.Add(label + "x", value.x.ToString());
-    }
-    if (!editFields.ContainsKey(label + "y")) {
-      editFields.Add(label + "y", value.y.ToString());
-    }
-    if (!editFields.ContainsKey(label + "z")) {
-      editFields.Add(label + "z", value.z.ToString());
-    }
-    GUILayout.BeginHorizontal();
-    GUILayout.Label(label + " : " + value + "   ");
-    editFields[label + "x"] = GUILayout.TextField(editFields[label + "x"], maxLenght);
-    editFields[label + "y"] = GUILayout.TextField(editFields[label + "y"], maxLenght);
-    editFields[label + "z"] = GUILayout.TextField(editFields[label + "z"], maxLenght);
-    if (GUILayout.Button(new GUIContent("Set", "Set vector"), GUILayout.Width(60f))) {
-      var tmpVector3 = new Vector3(float.Parse(editFields[label + "x"]),
-                                   float.Parse(editFields[label + "y"]),
-                                   float.Parse(editFields[label + "z"]));
-      value = tmpVector3;
-      btnPress = true;
-    }
-    GUILayout.EndHorizontal();
-    return btnPress;
-  }
-
-  public static bool EditField(string label, ref string value, int maxLenght = 50) {
-    bool btnPress = false;
-    if (!editFields.ContainsKey(label)) {
-      editFields.Add(label, value.ToString());
-    }
-    GUILayout.BeginHorizontal();
-    GUILayout.Label(label + " : " + value + "   ");
-    editFields[label] = GUILayout.TextField(editFields[label], maxLenght);
-    if (GUILayout.Button(new GUIContent("Set", "Set string"), GUILayout.Width(60f))) {
-      value = editFields[label];
-      btnPress = true;
-    }
-    GUILayout.EndHorizontal();
-    return btnPress;
-  }
-
-  public static bool EditField(string label, ref int value, int maxLenght = 50) {
-    bool btnPress = false;
-    if (!editFields.ContainsKey(label)) {
-      editFields.Add(label, value.ToString());
-    }
-    GUILayout.BeginHorizontal();
-    GUILayout.Label(label + " : " + value + "   ");
-    editFields[label] = GUILayout.TextField(editFields[label], maxLenght);
-    if (GUILayout.Button(new GUIContent("Set", "Set int"), GUILayout.Width(60f))) {
-      value = int.Parse(editFields[label]);
-      btnPress = true;
-    }
-    GUILayout.EndHorizontal();
-    return btnPress;
-  }
-
-  public static bool EditField(string label, ref float value, int maxLenght = 50) {
-    bool btnPress = false;
-    if (!editFields.ContainsKey(label)) {
-      editFields.Add(label, value.ToString());
-    }
-    GUILayout.BeginHorizontal();
-    GUILayout.Label(label + " : " + value + "   ");
-    editFields[label] = GUILayout.TextField(editFields[label], maxLenght);
-    if (GUILayout.Button(new GUIContent("Set", "Set float"), GUILayout.Width(60f))) {
-      value = float.Parse(editFields[label]);
-      btnPress = true;
-    }
-    GUILayout.EndHorizontal();
-    return btnPress;
   }
 
   /// <summary>Sets highlight status of the entire heierarchy.</summary>
@@ -951,14 +833,16 @@ public static class KIS_Shared {
   /// </remarks>
   /// <param name="assemblyRoot">Root of the assembly to move.</param>
   /// <param name="srcAttachNodeId">Attach node ID on the assembly root.</param>
-  /// <param name="tgtPart">New root part.</param>
+  /// <param name="tgtPart">
+  /// The part to align to. It can be <c>null</c> when dropping on the surface.
+  /// </param>
   /// <param name="tgtAttachNode">
   /// Attach node ID on the new target. It can be <c>null</c> if assembly is attached via surface
   /// node.
   /// </param>
   /// <param name="pos">Position of the assembly at the new parent.</param>
   /// <param name="rot">Rotation of the assembly at the new parent.</param>
-  /// <param name="onReady">Cllaback to execute when assembly move completed.</param>
+  /// <param name="onReady">Callback to execute when assembly move completed.</param>
   /// <returns>Enumerator that can be used as coroutine target.</returns>
   public static IEnumerator AsyncMoveAssembly(
       Part assemblyRoot, string srcAttachNodeId,
@@ -969,6 +853,15 @@ public static class KIS_Shared {
 
     var srcAttachNode = GetAttachNodeById(assemblyRoot, srcAttachNodeId);
     SendKISMessage(assemblyRoot, MessageAction.AttachStart, srcAttachNode, tgtPart, tgtAttachNode);
+
+    // Check if the target is a surface.
+    if (tgtPart == null) {
+      SendKISMessage(assemblyRoot, MessageAction.AttachEnd, srcAttachNode, tgtPart, tgtAttachNode);
+      if (onReady != null) {
+        onReady(assemblyRoot);
+      }
+      yield break;
+    }
 
     // Proactively disable collisions on the moving parts since there will be a period of time when
     // they don't belong to the target vessel.